Downgrade to v1 of vim-browse
I overwrote history a bit here because it was too complicated to
actually downgrade to v1 from v2.

Unfortunately while v2 is a nicer patch, it conflicts too much with the
mouse-scrollback patch for st. The eventual tipping points were:

1. If you're using vim-browse v2 you cannot use the mouse to scroll back
in terminal history past the first cut-off point in the output. This is
due to the `if historyScrollBuffer(n) return;` in `tscrollup`. Given I'm
not a C programmer, I could not figure out a way around that bit.

2. vim-browse v2 does not seem to play nicely with a command history
search tool I use with zsh (history-search-multi-word)

I'd imagine #2 is directly correlated with #1.

vim-browse v1 does not appear to have the same issues.